% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.frame.survfit.R
\name{as.data.frame.survfit}
\alias{as.data.frame.survfit}
\title{Convert a survfit to a dataframe of survival and censoring data}
\usage{
\method{as.data.frame}{survfit}(x)
}
\arguments{
\item{x}{survfit object (survival)}
}
\description{
This function returns a data frame containing following variables derived from a survfit object. Extends the S3 survfit class.
\itemize{
   \item t - Time
   \item s - KM Survival estimate
   \item c - Censoring indicator - 0 is censored, 1 is event
   \item strata - character description of strata that survival estimates are for
   \item Plus all strata variables
}
}
\examples{
library(survival)
sf <- survfit(Surv(time, status)~x, data = survival::aml)
as.data.frame(sf)
}
\keyword{survival}
