% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggatrisk.R
\name{ggatrisk}
\alias{ggatrisk}
\title{Plots at risk status from an rpsftm object}
\usage{
ggatrisk(x, eval.times = NULL)
}
\arguments{
\item{x}{rpsftm object (rpsftm)}

\item{eval.times}{- vector of times to evaluate at risk status for. Default is all event and censoring times.}
}
\description{
This function returns a ggplot derived from an rpsftm object showing status over time.
It assumes that only a single switch in treatment occurs i.e. a patient who is randomized to control but has 25% exposure to experimental is assumed to have switched at 0.75*event time.
}
\examples{
# use data included in rpsftm package
immdef <- rpsftm::immdef
immdef$rx <- with(immdef, 1 - xoyrs/progyrs)

# fit the model
imm.fit <- rpsftm::rpsftm(Surv(progyrs, prog) ~ rand(imm, rx),
                  data=immdef,
                  censor_time=censyrs)

ggatrisk(imm.fit)
}
\keyword{rpsftm}
\keyword{survival,}
