% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcfact.R
\name{ggcfact}
\alias{ggcfact}
\title{Plots observed and counterfactual survival from rpsftm object}
\usage{
ggcfact(x, incidence = FALSE)
}
\arguments{
\item{x}{rpsftm object (rpsftm)}

\item{incidence}{TRUE or FALSE (should survival or 1-survival be plotted)}
}
\description{
This function returns a ggplot derived from an rpsftm object showing 3 survival curves. Observed survival for experimental and control arm. Latent survival time for control arm.
}
\examples{
# use data included in rpsftm package
immdef <- rpsftm::immdef
immdef$rx <- with(immdef, 1 - xoyrs/progyrs)

# fit the model
imm.fit <- rpsftm::rpsftm(Surv(progyrs, prog) ~ rand(imm, rx),
                  data=immdef,
                  censor_time=censyrs)

ggcfact(imm.fit)
}
\keyword{rpsftm}
\keyword{survival,}
