% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdiscount.R
\name{ggdiscount}
\alias{ggdiscount}
\title{Plots hypothetical results from an rpsftm object}
\usage{
ggdiscount(x)
}
\arguments{
\item{x}{rpsftm object (rpsftm)}
}
\description{
This function returns a ggplot derived from an rpsftm object showing hazard ratio and 
95 percent CI for each psi tested in the g-estimation. A flat line suggests that there is no 
impact of switching on the results. A sloped line indicates that switching does affect the results.
}
\examples{
# use data included in rpsftm package
immdef <- rpsftm::immdef
immdef$rx <- with(immdef, 1 - xoyrs/progyrs)

# fit the model
imm.fit <- rpsftm::rpsftm(Surv(progyrs, prog) ~ rand(imm, rx),
                  data=immdef,
                  censor_time=censyrs)

ggdiscount(imm.fit)
}
\keyword{rpsftm}
\keyword{survival,}
