% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gglatent.R
\name{gglatent}
\alias{gglatent}
\title{Plots latent survival from rpsftm object}
\usage{
gglatent(x, incidence = FALSE)
}
\arguments{
\item{x}{rpsftm object (rpsftm)}

\item{incidence}{TRUE or FALSE (should survival or 1-survival be plotted)}
}
\description{
This function returns a ggplot derived from an rpsftm object showing latent survival time for experimental and control arms. If converged then the two latent survival curves should overlap.
}
\examples{
# use data included in rpsftm package
immdef <- rpsftm::immdef
immdef$rx <- with(immdef, 1 - xoyrs/progyrs)

# fit the model
imm.fit <- rpsftm(Surv(progyrs, prog) ~ rand(imm, rx),
                  data=immdef,
                  censor_time=censyrs)

gglatent(imm.fit)
}
\keyword{rpsftm}
\keyword{survival,}
