% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpsi.R
\name{ggpsi}
\alias{ggpsi}
\title{Plots g-estimation from rpsftm object}
\usage{
ggpsi(x)
}
\arguments{
\item{x}{rpsftm object (rpsftm)}
}
\description{
This function returns a ggplot derived from an rpsftm object showing psi vs test statistic. In case of convergence should see a unique solution.
}
\examples{
# use data included in rpsftm package
library(rpsftm)
immdef <- rpsftm::immdef
immdef$rx <- with(immdef, 1 - xoyrs/progyrs)

# fit the model
imm.fit <- rpsftm(Surv(progyrs, prog) ~ rand(imm, rx),
                  data=immdef,
                  censor_time=censyrs)

ggpsi(imm.fit)
}
\keyword{rpsftm}
\keyword{survival,}
