% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpsftm.atrisk.R
\name{rpsftm.atrisk}
\alias{rpsftm.atrisk}
\title{Evaluate at risk status from an rpsftm object}
\usage{
rpsftm.atrisk(x, eval.times = NULL)
}
\arguments{
\item{x}{rpsftm object (rpsftm)}

\item{eval.times}{- vector of times to evaluate at risk status for. Default is all event and censoring times.}
}
\description{
This function returns a data frame containing following variables derived from an rpsftm object.
It assumes that only a single switch in treatment occurs i.e. a patient who is randomized to control but has 25% exposure to experimental is assumed to have switched at 0.75*event time.
\itemize{
   \item .arm - 0 or 1 - randomized arm
   \item Arm - "Randomized to Control" or "Randomized to Experimental"
   \item Time - time that status applies for
   \item Status - one of "Event", "Censored", "At risk on Control", "At risk on Experimental"
   \item Count - count of patients in this status at this time.
   \item Percent - based on total number of patients in arm.
}
}
\examples{
# use data included in rpsftm package
library(rpsftm)
immdef <- rpsftm::immdef
immdef$rx <- with(immdef, 1 - xoyrs/progyrs)

# fit the model
imm.fit <- rpsftm(Surv(progyrs, prog) ~ rand(imm, rx),
                  data=immdef,
                  censor_time=censyrs)

rpsftm.atrisk(imm.fit, eval.times = c(0, 1, 2, 3))
}
\keyword{survival}
