% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpsftm.coxph.R
\name{rpsftm.coxph}
\alias{rpsftm.coxph}
\title{Fits a cox model to an rpsftm object}
\usage{
rpsftm.coxph(x)
}
\arguments{
\item{x}{rpsftm object (rpsftm)}
}
\description{
This function fits a cox model to an rpsftm object. It estimates a hazard ratio for the comparison of experimental observed time compared to control latent time. In addition the variance of the treatment coefficient is inflated based on the test used in g-estimation. Any strata included in rpsftm call are used in the cox model.
}
\examples{
# use data included in rpsftm package
library(rpsftm)
immdef <- rpsftm::immdef
immdef$rx <- with(immdef, 1 - xoyrs/progyrs)

# fit the model
imm.fit <- rpsftm(Surv(progyrs, prog) ~ rand(imm, rx),
                  data=immdef,
                  censor_time=censyrs)

rpsftm.coxph(imm.fit)
}
\keyword{rpsftm}
\keyword{survival,}
