% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpsftm.refit.R
\name{rpsftm.refit}
\alias{rpsftm.refit}
\title{Refits an rpsftm model for common treatment effect sensitivity analysis}
\usage{
rpsftm.refit(x, k)
}
\arguments{
\item{x}{rpsftm object (rpsftm)}

\item{k}{weight to be applied in the control arm for treatment effect. Single number.}
}
\description{
This function refits an rpsftm object with a specific weight applied to control arm efficacy. It is a short cut to recalling rpsftm with same parameters and varying the treat_modifier parameter.
}
\details{
See also documentation for rpsftm call treat_modifier= value.
}
\examples{
# use data included in rpsftm package
library(rpsftm)
immdef <- rpsftm::immdef
immdef$rx <- with(immdef, 1 - xoyrs/progyrs)

# fit the model
imm.fit <- rpsftm(Surv(progyrs, prog) ~ rand(imm, rx),
                  data=immdef,
                  censor_time=censyrs)

rpsftm.refit(imm.fit, k = 0.7)
rpsftm.refit(imm.fit, k = 1.3)
}
\keyword{rpsftm}
\keyword{survival,}
